rename the SearchPosts component as PostSearch to emphasize that the page shows posts only
